ARTEMIS-338  Refactor Journal Encodings into new package
