+++ artemis-server/src/main/java/org/apache/activemq/artemis/core/persistence/impl/journal/AckDescribe.java
@@ -14,40 +14,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.activemq.artemis.core.journal;
+package org.apache.activemq.artemis.core.persistence.impl.journal;
 
-import java.util.ArrayList;
-import java.util.List;
+import org.apache.activemq.artemis.core.persistence.impl.journal.codec.RefEncoding;
 
-public class PreparedTransactionInfo {
+public final class AckDescribe {
 
-   private final long id;
+   public RefEncoding refEncoding;
 
-   private final byte[] extraData;
-
-   private final List<RecordInfo> records = new ArrayList<RecordInfo>();
-
-   private final List<RecordInfo> recordsToDelete = new ArrayList<RecordInfo>();
-
-   public PreparedTransactionInfo(final long id, final byte[] extraData) {
-      this.id = id;
-
-      this.extraData = extraData;
+   public AckDescribe(RefEncoding refEncoding) {
+      this.refEncoding = refEncoding;
    }
 
-   public long getId() {
-      return id;
+   @Override
+   public String toString() {
+      return "ACK;" + refEncoding;
    }
 
-   public byte[] getExtraData() {
-      return extraData;
-   }
-
-   public List<RecordInfo> getRecords() {
-      return records;
-   }
-
-   public List<RecordInfo> getRecordsToDelete() {
-      return recordsToDelete;
-   }
 }
