+++ artemis-server/src/main/java/org/apache/activemq/artemis/core/persistence/impl/journal/AddMessageRecord.java
@@ -26,9 +26,11 @@ public final class AddMessageRecord {
 
    final ServerMessage message;
 
-   long scheduledDeliveryTime;
+   // mtaylor (Added to compile)
+   public long scheduledDeliveryTime;
 
-   int deliveryCount;
+   // mtaylor (Added to compile)
+   public int deliveryCount;
 
    public ServerMessage getMessage() {
       return message;
@@ -41,4 +43,5 @@ public final class AddMessageRecord {
    public int getDeliveryCount() {
       return deliveryCount;
    }
+
 }
