+++ artemis-server/src/main/java/org/apache/activemq/artemis/core/persistence/impl/journal/JournalRecordIds.java
@@ -26,7 +26,9 @@ package org.apache.activemq.artemis.core.persistence.impl.journal;
 public final class JournalRecordIds {
 
    // grouping journal record type
-   static final byte GROUP_RECORD = 20;
+
+   // mtaylor Added to compile
+   public static final byte GROUP_RECORD = 20;
 
    // BindingsImpl journal record type
