+++ artemis-server/src/main/java/org/apache/activemq/artemis/core/persistence/impl/journal/codec/QueueEncoding.java
@@ -14,40 +14,39 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.activemq.artemis.core.journal;
+package org.apache.activemq.artemis.core.persistence.impl.journal.codec;
 
-import java.util.ArrayList;
-import java.util.List;
+import org.apache.activemq.artemis.api.core.ActiveMQBuffer;
+import org.apache.activemq.artemis.core.journal.EncodingSupport;
 
-public class PreparedTransactionInfo {
+public class QueueEncoding implements EncodingSupport {
 
-   private final long id;
+   public long queueID;
 
-   private final byte[] extraData;
-
-   private final List<RecordInfo> records = new ArrayList<RecordInfo>();
-
-   private final List<RecordInfo> recordsToDelete = new ArrayList<RecordInfo>();
-
-   public PreparedTransactionInfo(final long id, final byte[] extraData) {
-      this.id = id;
-
-      this.extraData = extraData;
+   public QueueEncoding(final long queueID) {
+      super();
+      this.queueID = queueID;
    }
 
-   public long getId() {
-      return id;
+   public QueueEncoding() {
+      super();
    }
 
-   public byte[] getExtraData() {
-      return extraData;
+   public void decode(final ActiveMQBuffer buffer) {
+      queueID = buffer.readLong();
    }
 
-   public List<RecordInfo> getRecords() {
-      return records;
+   public void encode(final ActiveMQBuffer buffer) {
+      buffer.writeLong(queueID);
    }
 
-   public List<RecordInfo> getRecordsToDelete() {
-      return recordsToDelete;
+   public int getEncodeSize() {
+      return 8;
    }
+
+   @Override
+   public String toString() {
+      return "QueueEncoding [queueID=" + queueID + "]";
+   }
+
 }
