+++ artemis-server/src/main/java/org/apache/activemq/artemis/core/persistence/impl/journal/codec/ScheduledDeliveryEncoding.java
@@ -1,0 +1,54 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.activemq.artemis.core.persistence.impl.journal.codec;
+
+import org.apache.activemq.artemis.api.core.ActiveMQBuffer;
+
+public class ScheduledDeliveryEncoding extends QueueEncoding {
+
+   public long scheduledDeliveryTime;
+
+   @Override
+   public String toString() {
+      return "ScheduledDeliveryEncoding [scheduledDeliveryTime=" + scheduledDeliveryTime + "]";
+   }
+
+   public ScheduledDeliveryEncoding(final long scheduledDeliveryTime, final long queueID) {
+      super(queueID);
+      this.scheduledDeliveryTime = scheduledDeliveryTime;
+   }
+
+   public ScheduledDeliveryEncoding() {
+   }
+
+   @Override
+   public int getEncodeSize() {
+      return super.getEncodeSize() + 8;
+   }
+
+   @Override
+   public void encode(final ActiveMQBuffer buffer) {
+      super.encode(buffer);
+      buffer.writeLong(scheduledDeliveryTime);
+   }
+
+   @Override
+   public void decode(final ActiveMQBuffer buffer) {
+      super.decode(buffer);
+      scheduledDeliveryTime = buffer.readLong();
+   }
+}
